src/requirements.txt
python-dotenv~=0.20.0
requests~=2.25.1
munch~=2.5.0
twilio~=7.8.0
pytz~=2021.1
arrow~=1.2.2